chore(nested-flow): rename message to command
